// src/app/core/enums/widget-key.enum.ts
export enum WidgetsKey {
  ProductType = 'product-type',
  ShapeOfLabel = 'shape-of-label',
  Size = 'size',
  Material = 'material',
  Laminate = 'laminate',
  WhiteInk = 'white-ink',
  RoundedCorners = 'rounded-corners',
  Quantity = 'quantity',
  ApplicationMethod = 'application-method',
  LabelDirection = 'label-direction',
  Thickness = 'thickness',

  Width = 'width',
  Height = 'height',
  HandApplied = 'hand-applied',
  MachineApplied = 'machine-applied',
  Top = 'top',
  Bottom = 'bottom',
  Left = 'left',
  Right = 'right',
  GapBetweenLabels = 'gap-between-labels',
  GapBetweenLabelAndRollEdge = 'gap-between-label-and-roll-edge',
  LabelsPerRoll = 'labels-per-roll',
  MaxRollDiameter = 'max-roll-diameter',
  Kind = 'number-of-kinds',
  TotalQuantity = 'total-quantity',
  HandAappliedKey = 'hand-applied-application-method-application-method',
}